First cleanup round, reorganize/rename newly added methods, breaking some methods out into their own python modules.  Fix type annotations.
