+++ src/fidesops/service/privacy_request/request_runner_service.py
@@ -20,9 +20,9 @@ from fidesops.models.policy import (
 )
 from fidesops.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
 from fidesops.service.storage.storage_uploader_service import upload
+from fidesops.task.filter_results import filter_data_categories
 from fidesops.task.graph_task import (
     run_access_request,
-    filter_data_categories,
     run_erasure,
 )
 from fidesops.tasks.scheduled.scheduler import scheduler
