+++ tests/service/privacy_request/request_runner_service_test.py
@@ -1,5 +1,3 @@
-import json
-import logging
 import time
 from typing import Any, Dict, List, Set
 from unittest import mock
