+++ tests/task/test_graph_task.py
@@ -1,6 +1,4 @@
-import copy
 import pytest
-from datetime import datetime
 
 import dask
 from bson import ObjectId
@@ -17,7 +15,7 @@ from fidesops.task.graph_task import (
     collect_queries,
     TaskResources,
     EMPTY_REQUEST,
-    filter_data_categories,
+    consolidate_query_matches,
 )
 from fidesops.util.cache import get_cache
 from .traversal_data import sample_traversal, combined_mongo_postgresql_graph
@@ -297,505 +295,49 @@ def cache_input_helper(privacy_request_id, collection_address_str, value):
     cache.set_encoded_object(key, value)
 
 
-def test_filter_data_categories():
-    """Test different combinations of data categories to ensure the access_request_results are filtered properly"""
-    access_request_results = {
-        "postgres_example:supplies": [
-            {
-                "foods": {
-                    "vegetables": True,
-                    "fruits": {
-                        "apples": True,
-                        "oranges": False,
-                        "berries": {"strawberries": True, "blueberries": False},
-                    },
-                    "grains": {"rice": False, "wheat": True},
-                },
-                "clothing": True,
-            }
-        ]
-    }
+def test_consolidate_query_matches():
+    # Matching scalar returned
+    input_data = {"B": 55}
+    target_path = FieldPath("B")
+    assert consolidate_query_matches(input_data, target_path) == [55]
 
-    data_category_fields = {
-        CollectionAddress("postgres_example", "supplies"): {
-            "A": [FieldPath("foods", "fruits", "apples"), FieldPath("clothing")],
-            "B": [FieldPath("foods", "vegetables")],
-            "C": [
-                FieldPath("foods", "grains", "rice"),
-                FieldPath("foods", "grains", "wheat"),
-            ],
-            "D": [],
-            "E": [
-                FieldPath("foods", "fruits", "berries", "strawberries"),
-                FieldPath("foods", "fruits", "oranges"),
-            ],
-        }
-    }
+    # Matching array returned as-is
+    input_data = {"A": [1, 2, 3]}
+    target_path = FieldPath("A")
+    assert consolidate_query_matches(input_data, target_path) == [1, 2, 3]
 
-    only_a_categories = filter_data_categories(
-        copy.deepcopy(access_request_results), {"A"}, data_category_fields, "test_id"
-    )
+    # Array of embedded objects have multiple matching sub-paths merged
+    field_path = FieldPath("A", "B")
+    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
+    assert consolidate_query_matches(input_data, field_path) == [1, 3, 5]
 
-    assert only_a_categories == {
-        "postgres_example:supplies": [
-            {"foods": {"fruits": {"apples": True}}, "clothing": True}
-        ]
-    }
+    # Nested array returned
+    input_data = {"A": {"B": {"C": [9, 8, 7]}}, "D": {"E": {"F"}}}
+    field_path = FieldPath("A", "B", "C")
+    assert consolidate_query_matches(input_data, field_path) == [9, 8, 7]
 
-    only_b_categories = filter_data_categories(
-        copy.deepcopy(access_request_results), {"B"}, data_category_fields, "test_id"
-    )
-    assert only_b_categories == {
-        "postgres_example:supplies": [
-            {
-                "foods": {
-                    "vegetables": True,
-                }
-            }
-        ]
-    }
+    # Array of arrays are merged
+    input_data = {"A": [[5, 6], [7, 8], [9, 10]], "B": [[5, 6], [7, 8], [9, 10]]}
+    field_path = FieldPath("A")
+    assert consolidate_query_matches(input_data, field_path) == [5, 6, 7, 8, 9, 10]
 
-    only_c_categories = filter_data_categories(
-        copy.deepcopy(access_request_results), {"C"}, data_category_fields, "test_id"
-    )
-    assert only_c_categories == {
-        "postgres_example:supplies": [
-            {"foods": {"grains": {"rice": False, "wheat": True}}}
-        ]
-    }
-
-    only_d_categories = filter_data_categories(
-        copy.deepcopy(access_request_results), {"D"}, data_category_fields, "test_id"
-    )
-    assert only_d_categories == {}
-
-    only_e_categories = filter_data_categories(
-        copy.deepcopy(access_request_results), {"E"}, data_category_fields, "test_id"
-    )
-    assert only_e_categories == {
-        "postgres_example:supplies": [
-            {
-                "foods": {
-                    "fruits": {
-                        "oranges": False,
-                        "berries": {"strawberries": True},
-                    }
-                }
-            }
-        ]
-    }
-
-
-def test_filter_data_categories_arrays():
-    access_request_results = {
-        "postgres_example:flights": [
-            {
-                "people": {
-                    "passenger_ids": [222, 445, 311, 4444],
-                    "pilot_ids": [123, 12, 112],
-                },
-                "flight_number": 101,
-            }
-        ]
-    }
-
-    data_category_fields = {
-        CollectionAddress("postgres_example", "flights"): {
-            "A": [FieldPath("people", "passenger_ids")],
-            "B": [FieldPath("people", "pilot_ids")],
-        }
-    }
-
-    only_a_category = filter_data_categories(
-        copy.deepcopy(access_request_results), {"A"}, data_category_fields, "test_id"
-    )
-
-    # Nested array field retrieved
-    assert only_a_category == {
-        "postgres_example:flights": [
-            {"people": {"passenger_ids": [222, 445, 311, 4444]}}
-        ]
-    }
-
-    # Only matching cached value in nested array retrieved
-    cache_input_helper(
-        privacy_request_id="test_id",
-        collection_address_str="postgres_example:flights",
-        value={"people.pilot_ids": [112]},
-    )
-    only_b_category = filter_data_categories(
-        copy.deepcopy(access_request_results), {"B"}, data_category_fields, "test_id"
-    )
-    assert only_b_category == {
-        "postgres_example:flights": [{"people": {"pilot_ids": [112]}}]
-    }
-
-
-def test_filter_data_categories_limited_results():
-    """
-    Test scenario where the related data for a given identity is only a small subset of all the annotated fields.
-    """
-    jane_results = {
-        "mongo_test:customer_details": [
-            {
-                "_id": ObjectId("61f2bc8d6362fd78d72d8791"),
-                "customer_id": 3.0,
-                "gender": "female",
-                "birthday": datetime(1990, 2, 28, 0, 0),
-            }
+    # Array of arrays of embedded objects are merged
+    input_data = {
+        "A": [
+            [{"B": 1, "C": 2, "D": [3]}, {}],
+            [{"B": 3, "C": 4, "D": [5]}, {"B": 77, "C": 88, "D": [99]}],
         ],
-        "postgres_example:order_item": [],
-        "postgres_example:report": [],
-        "postgres_example:orders": [
-            {"customer_id": 3, "id": "ord_ddd-eee", "shipping_address_id": 4}
-        ],
-        "postgres_example:employee": [],
-        "postgres_example:address": [
-            {
-                "city": "Example Mountain",
-                "house": 1111,
-                "id": 4,
-                "state": "TX",
-                "street": "Example Place",
-                "zip": "54321",
-            }
-        ],
-        "postgres_example:visit": [],
-        "postgres_example:product": [],
-        "postgres_example:customer": [
-            {
-                "address_id": 4,
-                "created": datetime(2020, 4, 1, 11, 47, 42),
-                "email": "jane@example.com",
-                "id": 3,
-                "name": "Jane Customer",
-            }
-        ],
-        "postgres_example:service_request": [],
-        "postgres_example:payment_card": [
-            {
-                "billing_address_id": 4,
-                "ccn": 373719391,
-                "code": 222,
-                "customer_id": 3,
-                "id": "pay_ccc-ccc",
-                "name": "Example Card 3",
-                "preferred": False,
-            }
-        ],
-        "mongo_test:customer_feedback": [],
-        "postgres_example:login": [
-            {"customer_id": 3, "id": 8, "time": datetime(2021, 1, 6, 1, 0)}
-        ],
-        "mongo_test:internal_customer_profile": [],
+        "B": [[5, 6], [7, 8], [9, 10]],
     }
+    field_path = FieldPath("A", "D")
+    assert consolidate_query_matches(input_data, field_path) == [3, 5, 99]
 
-    target_categories = {"user.provided.identifiable"}
+    # Target path doesn't exist in data
+    field_path = FieldPath("A", "E", "X")
+    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
+    assert consolidate_query_matches(input_data, field_path) == []
 
-    data_category_fields = {
-        CollectionAddress.from_string("postgres_example:address"): {
-            "user.provided.identifiable.contact.city": [
-                FieldPath(
-                    "city",
-                )
-            ],
-            "user.provided.identifiable.contact.street": [
-                FieldPath(
-                    "house",
-                ),
-                FieldPath(
-                    "street",
-                ),
-            ],
-            "system.operations": [
-                FieldPath(
-                    "id",
-                )
-            ],
-            "user.provided.identifiable.contact.state": [
-                FieldPath(
-                    "state",
-                )
-            ],
-            "user.provided.identifiable.contact.postal_code": [
-                FieldPath(
-                    "zip",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:customer"): {
-            "system.operations": [
-                FieldPath(
-                    "address_id",
-                ),
-                FieldPath(
-                    "created",
-                ),
-            ],
-            "user.provided.identifiable.contact.email": [
-                FieldPath(
-                    "email",
-                )
-            ],
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "id",
-                )
-            ],
-            "user.provided.identifiable.name": [
-                FieldPath(
-                    "name",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:employee"): {
-            "system.operations": [
-                FieldPath(
-                    "address_id",
-                )
-            ],
-            "user.provided.identifiable.contact.email": [
-                FieldPath(
-                    "email",
-                )
-            ],
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "id",
-                )
-            ],
-            "user.provided.identifiable.name": [
-                FieldPath(
-                    "name",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:login"): {
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "customer_id",
-                )
-            ],
-            "system.operations": [
-                FieldPath(
-                    "id",
-                )
-            ],
-            "user.derived.nonidentifiable.sensor": [
-                FieldPath(
-                    "time",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:order_item"): {
-            "system.operations": [
-                FieldPath(
-                    "order_id",
-                ),
-                FieldPath(
-                    "product_id",
-                ),
-                FieldPath(
-                    "quantity",
-                ),
-            ]
-        },
-        CollectionAddress.from_string("postgres_example:orders"): {
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "customer_id",
-                )
-            ],
-            "system.operations": [
-                FieldPath(
-                    "id",
-                ),
-                FieldPath(
-                    "shipping_address_id",
-                ),
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:payment_card"): {
-            "system.operations": [
-                FieldPath(
-                    "billing_address_id",
-                ),
-                FieldPath(
-                    "id",
-                ),
-            ],
-            "user.provided.identifiable.financial.account_number": [
-                FieldPath(
-                    "ccn",
-                )
-            ],
-            "user.provided.identifiable.financial": [
-                FieldPath(
-                    "code",
-                ),
-                FieldPath(
-                    "name",
-                ),
-            ],
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "customer_id",
-                )
-            ],
-            "user.provided.nonidentifiable": [
-                FieldPath(
-                    "preferred",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:product"): {
-            "system.operations": [
-                FieldPath(
-                    "id",
-                ),
-                FieldPath(
-                    "name",
-                ),
-                FieldPath(
-                    "price",
-                ),
-            ]
-        },
-        CollectionAddress.from_string("postgres_example:report"): {
-            "user.provided.identifiable.contact.email": [
-                FieldPath(
-                    "email",
-                )
-            ],
-            "system.operations": [
-                FieldPath(
-                    "id",
-                ),
-                FieldPath(
-                    "month",
-                ),
-                FieldPath(
-                    "name",
-                ),
-                FieldPath(
-                    "total_visits",
-                ),
-                FieldPath(
-                    "year",
-                ),
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:service_request"): {
-            "user.provided.identifiable.contact.email": [
-                FieldPath(
-                    "alt_email",
-                )
-            ],
-            "system.operations": [
-                FieldPath(
-                    "closed",
-                ),
-                FieldPath(
-                    "email",
-                ),
-                FieldPath(
-                    "id",
-                ),
-                FieldPath(
-                    "opened",
-                ),
-            ],
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "employee_id",
-                )
-            ],
-        },
-        CollectionAddress.from_string("postgres_example:visit"): {
-            "user.provided.identifiable.contact.email": [
-                FieldPath(
-                    "email",
-                )
-            ],
-            "system.operations": [
-                FieldPath(
-                    "last_visit",
-                )
-            ],
-        },
-        CollectionAddress.from_string("mongo_test:customer_details"): {
-            "system.operations": [
-                FieldPath(
-                    "_id",
-                )
-            ],
-            "user.provided.identifiable.date_of_birth": [
-                FieldPath(
-                    "birthday",
-                )
-            ],
-            "user.derived.identifiable.unique_id": [
-                FieldPath(
-                    "customer_id",
-                )
-            ],
-            "user.provided.identifiable.gender": [
-                FieldPath(
-                    "gender",
-                )
-            ],
-            "user.provided.identifiable.job_title": [
-                FieldPath("workplace_info", "position")
-            ],
-        },
-        CollectionAddress.from_string("mongo_test:customer_feedback"): {
-            "system.operations": [
-                FieldPath(
-                    "_id",
-                )
-            ],
-            "user.provided.identifiable.contact.phone_number": [
-                FieldPath("customer_information", "phone")
-            ],
-            "user.provided.nonidentifiable": [
-                FieldPath(
-                    "message",
-                ),
-                FieldPath(
-                    "rating",
-                ),
-            ],
-        },
-        CollectionAddress.from_string("mongo_test:internal_customer_profile"): {
-            "user.derived": [
-                FieldPath(
-                    "derived_interests",
-                )
-            ]
-        },
-    }
-
-    filtered_results = filter_data_categories(
-        copy.deepcopy(jane_results), target_categories, data_category_fields, "test_id"
-    )
-
-    assert filtered_results == {
-        "mongo_test:customer_details": [
-            {"gender": "female", "birthday": datetime(1990, 2, 28)}
-        ],
-        "postgres_example:address": [
-            {
-                "city": "Example Mountain",
-                "house": 1111,
-                "state": "TX",
-                "street": "Example Place",
-                "zip": "54321",
-            }
-        ],
-        "postgres_example:customer": [
-            {"email": "jane@example.com", "name": "Jane Customer"}
-        ],
-        "postgres_example:payment_card": [
-            {"ccn": 373719391, "code": 222, "name": "Example Card 3"}
-        ],
-    }
+    # No field path
+    field_path = FieldPath()
+    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
+    assert consolidate_query_matches(input_data, field_path) == []
