--- tests/util/test_nested_utils.py
@@ -1,722 +1,0 @@
-import copy
-import pytest
-
-from fidesops.graph.config import FieldPath
-from fidesops.util.nested_utils import (
-    select_field_from_input_data,
-    flatten_and_merge_matches,
-    strip_empty_dicts,
-    refine_target_path,
-    _build_incoming_refined_target_paths,
-    _build_array_path_preservation,
-    remove_unmatched_array_paths,
-)
-
-
-def test_select_field():
-    final_results = {}
-    flat = {
-        "A": "a",
-        "B": "b",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {
-            "F": "g",
-            "H": "i",
-            "J": {"K": {"L": {"M": ["m", "n", "o"], "P": "p"}}, "N": {"O": "o"}},
-        },
-        "F": [{"G": "g", "H": "h"}, {"G": "h", "H": "i"}, {"G": "i", "H": "j"}],
-        "H": [
-            [
-                {"M": [1, 2, 3], "N": "n"},
-                {"M": [3, 2, 1], "N": "o"},
-                {"M": [1, 1, 1], "N": "p"},
-            ],
-            [
-                {"M": [4, 5, 6], "N": "q"},
-                {"M": [2, 2, 2], "N": "s"},
-                {"M": [], "N": "u"},
-            ],
-            [
-                {"M": [7, 8, 9], "N": "w"},
-                {"M": [6, 6, 6], "N": "y"},
-                {"M": [2], "N": "z"},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ],
-            "Y": [{"J": "l", "K": ["n"]}, {"J": "m", "K": ["customer@example.com"]}],
-            "Z": [{"J": "m", "K": ["n"]}],
-        },
-        "J": {
-            "K": {
-                "L": {
-                    "M": {
-                        "N": {
-                            "O": ["customer@example.com", "customer@gmail.com"],
-                            "P": ["customer@yahoo.com", "customer@customer.com"],
-                        }
-                    }
-                }
-            }
-        },
-        "K": [{"L": "l", "M": "m"}, {"L": "n", "M": "o"}],
-    }
-
-    # Test simple scalar field selected
-    assert select_field_from_input_data(final_results, flat, FieldPath("A")) == {
-        "A": "a"
-    }
-    # Test array field selected, and added to final results
-    assert select_field_from_input_data(final_results, flat, FieldPath("C")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-    }
-
-    # Test only certain scalar fields selected and added to final results
-    assert select_field_from_input_data(final_results, flat, FieldPath("D")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-    }
-    # Test nested field selected and added to final results
-    assert select_field_from_input_data(final_results, flat, FieldPath("E", "F")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g"},
-    }
-    # Test select field not in results
-    assert select_field_from_input_data(
-        final_results, flat, FieldPath("E", "F", "Z", "X")
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g"},
-    }
-
-    # Test more deeply nested scalar from previous dict
-    assert select_field_from_input_data(
-        final_results,
-        flat,
-        FieldPath("E", "J", "K", "L", "M"),
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-    }
-
-    # Test get matching dict key for each element in array
-    assert select_field_from_input_data(final_results, flat, FieldPath("F", "G")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-    }
-
-    import pdb
-
-    pdb.set_trace()
-    # Test get nested fields inside nested arrays
-    assert select_field_from_input_data(final_results, flat, FieldPath("H", "N")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [{"N": "n"}, {"N": "o"}, {"N": "p"}],
-            [{"N": "q"}, {"N": "s"}, {"N": "u"}],
-            [{"N": "w"}, {"N": "y"}, {"N": "z"}],
-        ],
-    }
-
-    import pdb
-
-    pdb.set_trace()
-    # Test get nested fields inside nested arrays
-    assert select_field_from_input_data(final_results, flat, FieldPath("H", "M")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-    }
-
-    # Test get dict of array of dict fields
-    assert select_field_from_input_data(
-        final_results, flat, FieldPath("I", "X", "J")
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {"X": [{"J": "j"}, {"J": "m"}]},
-    }
-
-    # Test get deeply nested array field with only matching data, array in arrays
-    assert select_field_from_input_data(
-        final_results, flat, FieldPath("I", "X", "K")
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ]
-        },
-    }
-
-    # Get deeply nested array inside of dicts, with only matching data
-    assert select_field_from_input_data(
-        final_results, flat, FieldPath("I", "Y", "K")
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ],
-            "Y": [{"K": ["n"]}, {"K": ["customer@example.com"]}],
-        },
-    }
-
-    assert select_field_from_input_data(
-        final_results,
-        flat,
-        FieldPath("J", "K", "L", "M", "N", "O"),
-    ) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ],
-            "Y": [{"K": ["n"]}, {"K": ["customer@example.com"]}],
-        },
-        "J": {
-            "K": {
-                "L": {"M": {"N": {"O": ["customer@example.com", "customer@gmail.com"]}}}
-            }
-        },
-    }
-
-    # Test "only" param does not apply to regular scalar fields
-    assert select_field_from_input_data(final_results, flat, FieldPath("B")) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ],
-            "Y": [{"K": ["n"]}, {"K": ["customer@example.com"]}],
-        },
-        "J": {
-            "K": {
-                "L": {"M": {"N": {"O": ["customer@example.com", "customer@gmail.com"]}}}
-            }
-        },
-        "B": "b",
-    }
-
-    assert select_field_from_input_data(final_results, flat, FieldPath("K", "L"),) == {
-        "A": "a",
-        "C": ["d", "e", "f"],
-        "D": ["g", "h", "i", "j"],
-        "E": {"F": "g", "J": {"K": {"L": {"M": ["m", "n", "o"]}}}},
-        "F": [{"G": "g"}, {"G": "h"}, {"G": "i"}],
-        "H": [
-            [
-                {"N": "n", "M": [1, 2, 3]},
-                {"N": "o", "M": [3, 2, 1]},
-                {"N": "p", "M": [1, 1, 1]},
-            ],
-            [
-                {"N": "q", "M": [4, 5, 6]},
-                {"N": "s", "M": [2, 2, 2]},
-                {"N": "u", "M": []},
-            ],
-            [
-                {"N": "w", "M": [7, 8, 9]},
-                {"N": "y", "M": [6, 6, 6]},
-                {"N": "z", "M": [2]},
-            ],
-        ],
-        "I": {
-            "X": [
-                {"J": "j", "K": ["k"]},
-                {"J": "m", "K": ["customer@example.com", "customer-1@example.com"]},
-            ],
-            "Y": [{"K": ["n"]}, {"K": ["customer@example.com"]}],
-        },
-        "J": {
-            "K": {
-                "L": {"M": {"N": {"O": ["customer@example.com", "customer@gmail.com"]}}}
-            }
-        },
-        "B": "b",
-        "K": [{"L": "l"}, {"L": "n"}],
-    }
-
-
-def test_flatten_and_merge_matches():
-    # Matching scalar returned
-    input_data = {"B": 55}
-    target_path = FieldPath("B")
-    assert flatten_and_merge_matches(input_data, target_path) == [55]
-
-    # Matching array returned as-is
-    input_data = {"A": [1, 2, 3]}
-    target_path = FieldPath("A")
-    assert flatten_and_merge_matches(input_data, target_path) == [1, 2, 3]
-
-    # Array of dicts have multiple matching sub-paths merged
-    field_path = FieldPath("A", "B")
-    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
-    assert flatten_and_merge_matches(input_data, field_path) == [1, 3, 5]
-
-    # Nested array returned
-    input_data = {"A": {"B": {"C": [9, 8, 7]}}, "D": {"E": {"F"}}}
-    field_path = FieldPath("A", "B", "C")
-    assert flatten_and_merge_matches(input_data, field_path) == [9, 8, 7]
-
-    # Array of arrays are merged
-    input_data = {"A": [[5, 6], [7, 8], [9, 10]], "B": [[5, 6], [7, 8], [9, 10]]}
-    field_path = FieldPath("A")
-    assert flatten_and_merge_matches(input_data, field_path) == [5, 6, 7, 8, 9, 10]
-
-    # Array of arrays of dicts are merged
-    input_data = {
-        "A": [
-            [{"B": 1, "C": 2, "D": [3]}, {}],
-            [{"B": 3, "C": 4, "D": [5]}, {"B": 77, "C": 88, "D": [99]}],
-        ],
-        "B": [[5, 6], [7, 8], [9, 10]],
-    }
-    field_path = FieldPath("A", "D")
-    assert flatten_and_merge_matches(input_data, field_path) == [3, 5, 99]
-
-    # Target path doesn't exist in data
-    field_path = FieldPath("A", "E", "X")
-    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
-    assert flatten_and_merge_matches(input_data, field_path) == []
-
-    # No field path
-    field_path = FieldPath()
-    input_data = {"A": [{"B": 1, "C": 2}, {"B": 3, "C": 4}, {"B": 5, "C": 6}]}
-    assert flatten_and_merge_matches(input_data, field_path) == []
-
-
-def test_strip_empty_dicts():
-    # No empty dicts, strip_empty_dicts has no effect
-    orig = {"A": {"B": {"C": []}, "G": {"H": None}}, "I": 0, "J": False}
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == orig
-
-    # Deeply nested dicts with no values removed - G - H - I levels gone.
-    orig = {"A": {"B": {"C": []}, "G": {"H": {"I": {}}}}, "I": 0}
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == {"A": {"B": {"C": []}}, "I": 0}
-
-    orig = {
-        "A": [{"B": "C", "D": {}}, {"B": "G", "D": {}}, {"B": "J", "D": {"J": "K"}}]
-    }
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == {"A": [{"B": "C"}, {"B": "G"}, {"B": "J", "D": {"J": "K"}}]}
-
-    # Empty dict returns original empty dict
-    orig = {}
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == {}
-
-    # Empty dict returned
-    orig = {"A": {}}
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == {}
-
-    # Empty dict *not* removed entirely from array - unclear what expected behavior
-    # is here. Original version was removing it, but I think that may be too aggressive.
-    orig = {"A": [[{"B": "C", "D": [{"F": {}}, {"G": []}]}, {"B": "D"}, {"B": "G"}]]}
-    results = copy.deepcopy(orig)
-    strip_empty_dicts(results)
-    assert results == {
-        "A": [[{"B": "C", "D": [{}, {"G": []}]}, {"B": "D"}, {"B": "G"}]]
-    }
-
-
-@pytest.fixture(scope="function")
-def sample_data():
-    return {
-        "_id": 12345,
-        "thread": [
-            {
-                "comment": "com_0001",
-                "message": "hello, testing in-flight chat feature",
-                "chat_name": "John",
-                "messages": {},
-            },
-            {
-                "comment": "com_0002",
-                "message": "yep, got your message, looks like it works",
-                "chat_name": "Jane",
-            },
-            {"comment": "com_0002", "message": "hello!", "chat_name": "Jeanne"},
-        ],
-        "snacks": ["pizza", "chips"],
-        "seats": {"first_choice": "A2", "second_choice": "B3"},
-        "upgrades": {
-            "magazines": ["Time", "People"],
-            "books": ["Once upon a Time", "SICP"],
-            "earplugs": True,
-        },
-        "other_flights": [
-            {"DFW": ["11 AM", "12 PM"], "CHO": ["12 PM", "1 PM"]},
-            {"DFW": ["2 AM", "12 PM"], "CHO": ["2 PM", "1 PM"]},
-            {"DFW": ["3 AM", "2 AM"], "CHO": ["2 PM", "1:30 PM"]},
-        ],
-        "months": {
-            "july": [
-                {
-                    "activities": ["swimming", "hiking"],
-                    "crops": ["watermelon", "cheese", "grapes"],
-                },
-                {"activities": ["tubing"], "crops": ["corn"]},
-            ],
-            "march": [
-                {
-                    "activities": ["skiing", "bobsledding"],
-                    "crops": ["swiss chard", "swiss chard"],
-                },
-                {"activities": ["hiking"], "crops": ["spinach"]},
-            ],
-        },
-        "hello": [1, 2, 3, 4, 2],
-        "weights": [[1, 2], [3, 4]],
-        "toppings": [[["pepperoni", "salami"], ["pepperoni", "cheese", "cheese"]]],
-        "A": {"C": [{"M": ["p", "n", "n"]}]},
-        "C": [["A", "B", "C", "B"], ["G", "H", "B", "B"]],  # Double lists
-        "D": [
-            [["A", "B", "C", "B"], ["G", "H", "B", "B"]],
-            [["A", "B", "C", "B"], ["G", "H", "B", "B"]],
-        ],  # Triple lists
-        "E": [[["B"], [["A", "B", "C", "B"], ["G", "H", "B", "B"]]]],  # Irregular lists
-        "F": [
-            "a",
-            ["1", "a", [["z", "a", "a"]]],
-        ],  # Lists elems are different types, not officially supported
-    }
-
-
-def test_refine_target_path(sample_data):
-    result = refine_target_path(
-        sample_data, ["months", "march", "crops"], ["swiss chard"]
-    )
-    assert result == [
-        ["months", "march", 0, "crops", 0],
-        ["months", "march", 0, "crops", 1],
-    ]
-
-    result = refine_target_path(sample_data, ["_id"], [12345])
-    assert result == ["_id"]
-
-    result = refine_target_path(sample_data, ["snacks"], ["pizza"])
-    assert result == ["snacks", 0]
-
-    result = refine_target_path(sample_data, ["thread", "comment"], ["com_0002"])
-    assert result == [["thread", 1, "comment"], ["thread", 2, "comment"]]
-
-    result = refine_target_path(sample_data, ["seats", "first_choice"], ["A2"])
-    assert result == ["seats", "first_choice"]
-
-    result = refine_target_path(sample_data, ["upgrades", "books"], ["SICP"])
-    assert result == ["upgrades", "books", 1]
-
-    result = refine_target_path(sample_data, ["other_flights", "CHO"], ["1 PM"])
-    assert result == [["other_flights", 0, "CHO", 1], ["other_flights", 1, "CHO", 1]]
-
-    result = refine_target_path(sample_data, ["bad_path"], ["bad match"])
-    assert result == []
-
-    result = refine_target_path(sample_data, ["hello"], only=[2])
-    assert result == [["hello", 1], ["hello", 4]]
-
-    result = refine_target_path(
-        sample_data, ["months", "july", "crops"], ["watermelon", "grapes"]
-    )
-    assert result == [
-        ["months", "july", 0, "crops", 0],
-        ["months", "july", 0, "crops", 2],
-    ]
-
-    result = refine_target_path(sample_data, ["weights"], [4])
-    assert result == ["weights", 1, 1]
-
-    result = refine_target_path(sample_data, ["toppings"], ["cheese"])
-    assert result == [["toppings", 0, 1, 1], ["toppings", 0, 1, 2]]
-
-    result = refine_target_path(sample_data, ["A", "C", "M"], ["n"])
-    assert result == [["A", "C", 0, "M", 1], ["A", "C", 0, "M", 2]]
-
-    result = refine_target_path(sample_data, [], ["pizza"])
-    assert result == []
-
-    result = refine_target_path(sample_data, ["C"], ["B"])
-    assert result == [["C", 0, 1], ["C", 0, 3], ["C", 1, 2], ["C", 1, 3]]
-
-    result = refine_target_path(sample_data, ["D"], ["B"])
-    assert result == [
-        ["D", 0, 0, 1],
-        ["D", 0, 0, 3],
-        ["D", 0, 1, 2],
-        ["D", 0, 1, 3],
-        ["D", 1, 0, 1],
-        ["D", 1, 0, 3],
-        ["D", 1, 1, 2],
-        ["D", 1, 1, 3],
-    ]
-
-    result = refine_target_path(sample_data, ["E"], ["B"])
-    assert result == [
-        ["E", 0, 0, 0],
-        ["E", 0, 1, 0, 1],
-        ["E", 0, 1, 0, 3],
-        ["E", 0, 1, 1, 2],
-        ["E", 0, 1, 1, 3],
-    ]
-
-    result = refine_target_path(sample_data, ["F"], ["a"])
-    assert result == [["F", 0], ["F", 1, 1], ["F", 1, 2, 0, 1], ["F", 1, 2, 0, 2]]
-
-
-def test_build_incoming_refined_target_paths(sample_data):
-    incoming_paths = {
-        FieldPath(
-            "F",
-        ): ["a"],
-        FieldPath("snacks"): ["pizza"],
-        FieldPath("thread", "comment"): ["com_0002"],
-    }
-    result = _build_incoming_refined_target_paths(sample_data, incoming_paths)
-    assert result == [
-        ["F", 0],
-        ["snacks", 0],
-        ["F", 1, 1],
-        ["thread", 1, "comment"],
-        ["thread", 2, "comment"],
-        ["F", 1, 2, 0, 1],
-        ["F", 1, 2, 0, 2],
-    ]
-
-
-def test_build_array_path_preservation():
-    expanded_field_paths = [
-        ["F", 0],
-        ["snacks", 0],
-        ["F", 1, 1],
-        ["thread", 1, "comment"],
-        ["thread", 2, "comment"],
-        ["F", 1, 2, 0, 1],
-        ["F", 1, 2, 0, 2],
-        ["Non", "integer"],
-    ]
-
-    assert _build_array_path_preservation(expanded_field_paths) == {
-        "F": [0, 1],
-        "snacks": [0],
-        "F.1": [1, 2],
-        "thread": [1, 2],
-        "F.1.2": [0],
-        "F.1.2.0": [1, 2],
-    }
-
-
-def test_removed_unmatched_array_paths(sample_data):
-    incoming_paths = {
-        FieldPath(
-            "F",
-        ): ["a"],
-        FieldPath("snacks"): ["pizza"],
-        FieldPath("thread", "comment"): ["com_0002"],
-    }
-
-    filtered_record = remove_unmatched_array_paths(sample_data, incoming_paths)
-    assert filtered_record == {
-        "_id": 12345,
-        "thread": [
-            {
-                "comment": "com_0002",
-                "message": "yep, got your message, looks like it works",
-                "chat_name": "Jane",
-            },
-            {"comment": "com_0002", "message": "hello!", "chat_name": "Jeanne"},
-        ],
-        "snacks": ["pizza"],
-        "seats": {"first_choice": "A2", "second_choice": "B3"},
-        "upgrades": {
-            "magazines": ["Time", "People"],
-            "books": ["Once upon a Time", "SICP"],
-            "earplugs": True,
-        },
-        "other_flights": [
-            {"DFW": ["11 AM", "12 PM"], "CHO": ["12 PM", "1 PM"]},
-            {"DFW": ["2 AM", "12 PM"], "CHO": ["2 PM", "1 PM"]},
-            {"DFW": ["3 AM", "2 AM"], "CHO": ["2 PM", "1:30 PM"]},
-        ],
-        "months": {
-            "july": [
-                {
-                    "activities": ["swimming", "hiking"],
-                    "crops": ["watermelon", "cheese", "grapes"],
-                },
-                {"activities": ["tubing"], "crops": ["corn"]},
-            ],
-            "march": [
-                {
-                    "activities": ["skiing", "bobsledding"],
-                    "crops": ["swiss chard", "swiss chard"],
-                },
-                {"activities": ["hiking"], "crops": ["spinach"]},
-            ],
-        },
-        "hello": [1, 2, 3, 4, 2],
-        "weights": [[1, 2], [3, 4]],
-        "toppings": [[["pepperoni", "salami"], ["pepperoni", "cheese", "cheese"]]],
-        "A": {"C": [{"M": ["p", "n", "n"]}]},
-        "C": [["A", "B", "C", "B"], ["G", "H", "B", "B"]],
-        "D": [
-            [["A", "B", "C", "B"], ["G", "H", "B", "B"]],
-            [["A", "B", "C", "B"], ["G", "H", "B", "B"]],
-        ],
-        "E": [[["B"], [["A", "B", "C", "B"], ["G", "H", "B", "B"]]]],
-        "F": ["a", ["a", [["a", "a"]]]],
-    }
